move back the tests to avoid conflicts with other PRs
